backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
google-cloud-firestore>=2.14.0
google-cloud-storage>=2.14.0
google-cloud-aiplatform>=1.38.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
